+++ test/helper.rb
@@ -19,8 +19,6 @@ require 'minitest/autorun'
 require 'minitest/mock'
 require_relative 'helpers/memcached'
 
-ENV['SASL_CONF_PATH'] = "#{File.dirname(__FILE__)}/sasl/memcached.conf"
-
 require 'dalli'
 require 'logger'
 require 'securerandom'