+++ test/protocol/test_request_formatter.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../helper'
+require_relative '../helper'
 
 describe Dalli::Protocol::Meta::RequestFormatter do
   describe 'meta_get' do