+++ test/protocol/test_response_processor.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../helper'
+require_relative '../helper'
 
 describe Dalli::Protocol::Meta::ResponseProcessor do
   let(:io_source) { Minitest::Mock.new }